+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## unreleased
 
+## [2024.12.3] - 2024-12-24
+
 ### Added
 
 - Render activity, attendance and savings for each subscription.
@@ -236,8 +238,9 @@ Initial release with the following features:
 - services like opening the gate,
 - user attendance history.
 
-[unreleased]: https://github.com/coworking-metz/mobile-app/compare/2024.12.2...main
-[2024.12.2: https://github.com/coworking-metz/mobile-app/compare/2024.12.1...2024.12.2
+[unreleased]: https://github.com/coworking-metz/mobile-app/compare/2024.12.3...main
+[2024.12.3]: https://github.com/coworking-metz/mobile-app/compare/2024.12.2...2024.12.3
+[2024.12.2]: https://github.com/coworking-metz/mobile-app/compare/2024.12.1...2024.12.2
 [2024.12.1]: https://github.com/coworking-metz/mobile-app/compare/2024.9.2...2024.12.1
 [2024.9.2]: https://github.com/coworking-metz/mobile-app/compare/2024.9.1...2024.9.2
 [2024.9.1]: https://github.com/coworking-metz/mobile-app/compare/2024.5.2...2024.9.1